refatoração extensão das injeçoes de dependências.
