+++ TechChallange.Api/Program.cs
@@ -12,6 +12,7 @@ using System.Diagnostics;
 using System.Text;
 using TechChallangeApi.Controllers;
 using TechChallangeApi.Data;
+using TechChallenge.Api.Extensions;
 using TechChallenge.Api.Factory;
 using TechChallenge.Api.Models;
 using TechChallenge.Api.Service;
@@ -20,46 +21,13 @@ var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 
-//Log configuration
-var configuration = new ConfigurationBuilder()
-	.SetBasePath(Directory.GetCurrentDirectory())
-			.AddJsonFile("appsettings.json")
-			.Build();
-
-var connectionStringLog = builder.Configuration.GetConnectionString("Log_SQL_CONNECTIONSTRING");
-
-var option = new ColumnOptions
-{
-	AdditionalColumns = new Collection<SqlColumn>
-	{
-		new SqlColumn {ColumnName = "Action"}
-	}
-};
-
-var sinkOpts = new MSSqlServerSinkOptions();
-
-Serilog.Log.Logger = new LoggerConfiguration()
-			.ReadFrom.Configuration(configuration)
-			.CreateLogger();
-
-Serilog.Debugging.SelfLog.Enable(msg =>
-{
-	Debug.Print(msg);
-	Debugger.Break();
-});
-
-builder.Host.UseSerilog();
-
+builder.Services.AddDILogs(builder);
 
 //Configurações DB
 var connectionString = builder.Configuration.GetConnectionString("AZURE_SQL_CONNECTIONSTRING");
 builder.Services.AddDbContext<ApplicationContext>(opts => opts.UseSqlServer(connectionString));
 
-//Add Services
-builder.Services.AddScoped<IFotosRepository, FotosRepository>();
-builder.Services.AddScoped<IFotosRepository, FotosRepository>();
-builder.Services.AddScoped<IPublicacaoRepository, PublicacaoRepository>();
-builder.Services.AddScoped<IUsuarioRepository, UsuarioRepository>();
+builder.Services.AddDIServices(builder);
 
 //Add Facotry
 builder.Services.AddScoped<ILogFactory, LogFactory>();
@@ -74,35 +42,9 @@ builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
-//cors
-var politica = "CorsPolicy-public";
+builder.Services.AddDICors(builder);
 
-builder.Services.AddCors(option => option.AddPolicy(politica, builder => builder.WithOrigins("http://localhost:4200", "https://localhost")
-	 .AllowAnyMethod()
-				.AllowAnyHeader()
-				.AllowCredentials()
-				.Build()));
-
-//authentication
-var key = Encoding.ASCII.GetBytes("1f1ce09a-0b07-4fd8-889e-e3e18442b081");
-
-builder.Services.AddAuthentication(x =>
-{
-	x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-	x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-}).AddJwtBearer(x =>
-{
-	x.RequireHttpsMetadata = false;
-	x.SaveToken = true;
-	x.TokenValidationParameters = new TokenValidationParameters
-	{
-		ValidateIssuerSigningKey = true,
-		IssuerSigningKey = new SymmetricSecurityKey(key),
-		ValidateIssuer = false,
-		ValidateAudience = false
-	};
-});
-
+builder.Services.AddDIAuthentication(builder);
 
 var app = builder.Build();
